.history/requirements_20241026211526.txt
PyQt5==5.15.4
Pillow==9.2.0
pillow-heif==0.7.0
piexif==1.1.3
